finalmente

+++ src/ArvoreBinaria/No.java
@@ -2,7 +2,7 @@ package ArvoreBinaria;
 
 
 public class No {
-    public long elemento;
+    public int elemento;
     public No direita;
     public No esquerda;
 }
